server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
h11==0.14.0
idna==3.4
numpy==1.26.0
opencv-python==4.8.1.78
pandas==2.1.1
Pillow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
wheel==0.41.2